+++ src/modules/post/section-1/section-1c/section1c.formatter.ts
