+++ src/modules/post/section-4/chicken-farm/chicken-farm.formatter.ts
