+++ src/modules/post/section-4/manufacture/manufacture.scene.ts
